Refactored location where download is started
I came across the issue of progressbar2 not detecting the terminal width correctly.
I correctly assumed it was because ServiceListener was running in some sort of separate thread. I figured it wasn't a good idea to have the download logic in there anyways so I refactored it to only report the address and port back to the main thread and start the download there.